# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5e1efcb50cfc6233ae931a2f
  title: MyUber
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []   # No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:   # Sample object
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: [] #No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        '200':
          description: OK
                

  #-----  Booking  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    User:
      x-skaffolder-id: 5e1efcb50cfc6233ae931a39
      x-skaffolder-id-db: 5e1efcb50cfc6233ae931a30
      x-skaffolder-id-entity: 5e1efcb50cfc6233ae931a31
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5e1efcb50cfc6233ae931a34
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5e1efcb50cfc6233ae931a35
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5e1efcb50cfc6233ae931a33
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5e1efcb50cfc6233ae931a37
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5e1efcb50cfc6233ae931a36
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5e1efcb50cfc6233ae931a32
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
      
    
    
  # Databases
...